added comment and reply upvote and some validation

+++ src/Firebase/api/database/PostsApi.ts
@@ -17,7 +17,7 @@ import {
 
 import { db } from "./database";
 import { auth } from "../auth/auth";
-import { ApiPost, Post } from "../../../Types/Posts";
+import { ApiPost, Post, postSchema } from "../../../Types/Posts";
 import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
 import { storage } from "./database";
 
@@ -55,9 +55,12 @@ export async function getAllPosts(): Promise<ApiPost[]> {
  * Retrieves the posts of a user from the API.
  *
  * @param {string} userId - The ID of the user whose posts will be retrieved.
+ * @throws {Error} If the user ID is empty
  * @return {Promise<ApiPost[]>} A promise that resolves to an array of ApiPost objects representing the user's posts.
  */
 export async function getUserPosts(userId: string): Promise<ApiPost[]> {
+  if (!userId) throw new Error("userId cannot be empty");
+
   const userData = (await getDoc(doc(db, "users", userId))).data();
   const q = query(postsCollectionRef, where("user_id", "==", userId));
   const querySnapshot = await getDocs(q);
@@ -74,9 +77,12 @@ export async function getUserPosts(userId: string): Promise<ApiPost[]> {
  * Retrieves a post from the API by its ID.
  *
  * @param {string} id - The ID of the post.
+ * @throws {Error} If the ID is empty
  * @return {Promise<ApiPost>} A promise that resolves to the retrieved post.
  */
 export async function getPostById(id: string): Promise<ApiPost> {
+  if (!id) throw new Error("id cannot be empty");
+
   const docRef = doc(db, "posts", id);
   const docSnap = await getDoc(docRef);
   const docData = docSnap.data();
@@ -92,9 +98,14 @@ export async function getPostById(id: string): Promise<ApiPost> {
  * Creates a new post.
  *
  * @param {Post} post - The post object containing the details of the post.
+ * @throws {Error} If post is not a valid Post object
+ * @throws {Error} If post is empty
  * @return {Promise<string>} A Promise that resolves to the ID of the newly created post.
  */
 export async function createPost(post: Post): Promise<string> {
+  if (!post) throw new Error("post cannot be empty");
+  if (postSchema.safeParse(post).success === false)
+    throw new Error("post must be a valid Post object");
   console.log("post from api", post);
 
   const newPostId = await addDoc(postsCollectionRef, {
@@ -136,21 +147,28 @@ export async function createPost(post: Post): Promise<string> {
  * Creates a new post.
  *
  * @param {Post} post - The post object containing the details of the post.
+ * @throws {Error} If the id is empty.
  * @return {Promise<string>} A Promise that resolves to the ID of the newly created post.
  */
 export async function deletePostById(id: string): Promise<void> {
+  if (!id) throw new Error("id cannot be empty");
+
   const docRef = doc(db, "posts", id);
   return await deleteDoc(docRef);
 }
 
 /**
- * Vote on a post.
+ * Updates the vote count for a post.
  *
  * @param {string} id - The ID of the post.
- * @param {"up" | "down"} vote - The type of vote ("up" or "down").
- * @return {Promise<void>} Promise that resolves when the vote is successfully recorded.
+ * @param {"up" | "down"} vote - The vote type ("up" or "down").
+ * @throws {Error} If the vote is not "up" or "down".
+ * @return {Promise<void>} - A promise that resolves when the vote is updated.
  */
 export async function votePost(id: string, vote: "up" | "down"): Promise<void> {
+  if (vote !== "up" && vote !== "down")
+    throw new Error(`Invalid vote, vote should only be "up" or "down"`);
+
   const docRef = doc(db, "posts", id);
   const docData = (await getDoc(docRef)).data();
